Move Rasterizer class
