+++ reVX/exclusions/setbacks/base.py
@@ -4,10 +4,7 @@ Base classes for setback exclusion computation
 """
 import os
 import logging
-from copy import deepcopy
 from warnings import warn
-from math import floor, ceil
-from itertools import product
 from abc import abstractmethod
 from concurrent.futures import as_completed
 
@@ -15,14 +12,12 @@ import numpy as np
 import geopandas as gpd
 from shapely.ops import unary_union
 from shapely.validation import make_valid
-from rasterio import (windows, transform, Affine, coords,
-                      features as rio_features)
+from rasterio import transform, coords
 
 from rex.utilities import log_mem
 from rex.utilities.execution import SpawnProcessPool
-from reV.handlers.exclusions import ExclusionLayers
 from reVX.handlers.geopackage import GPKGMeta
-from reVX.exclusions.base import AbstractBaseExclusionsMerger
+from reVX.exclusions.base import AbstractBaseExclusionsMerger, Rasterizer
 from reVX.exclusions.setbacks.functions import (
     parcel_buffer, positive_buffer, features_clipped_to_county,
     features_with_centroid_in_county)
@@ -42,186 +37,6 @@ FEATURE_FILTERS = {
     "clip": features_clipped_to_county,
 }
 """Types of feature filters available for setback calculations. """
-
-
-class Rasterizer:
-    """Helper class to rasterize shapes."""
-
-    def __init__(self, excl_fpath, weights_calculation_upscale_factor,
-                 hsds=False):
-        """
-
-        Parameters
-        ----------
-        excl_fpath : str
-            Path to .h5 file containing template layers. The raster will
-            match the shape and profile of these layers.
-        weights_calculation_upscale_factor : int
-            If this value is an int > 1, the output will be a layer with
-            **inclusion** weight values (floats ranging from 0 to 1).
-            Note that this is backwards w.r.t the typical output of
-            exclusion integer values (1 for excluded, 0 otherwise).
-            Values <= 1 will still return a standard exclusion mask.
-            For example, a cell that was previously excluded with a
-            a boolean mask (value of 1) may instead be converted to an
-            inclusion weight value of 0.75, meaning that 75% of the area
-            corresponding to that point should be included (i.e. the
-            exclusion feature only intersected a small portion - 25% -
-            of the cell). This percentage inclusion value is calculated
-            by upscaling the output array using this input value,
-            rasterizing the exclusion features onto it, and counting the
-            number of resulting sub-cells excluded by the feature. For
-            example, setting the value to 3 would split each output
-            cell into nine sub-cells - 3 divisions in each dimension.
-            After the feature is rasterized on this high-resolution
-            sub-grid, the area of the non-excluded sub-cells is totaled
-            and divided by the area of the original cell to obtain the
-            final inclusion percentage. Therefore, a larger upscale
-            factor results in more accurate percentage values. If
-            ``None`` (or a value <= 1), this process is skipped and the
-            output is a boolean exclusion mask. By default ``None``.
-        """
-        props = _parse_excl_properties(excl_fpath, hsds=hsds)
-        self._shape, self._profile = props
-        self._scale_factor = weights_calculation_upscale_factor or 1
-        self._scale_factor = int(self._scale_factor // 1)
-
-    @property
-    def profile(self):
-        """Geotiff profile.
-
-        Returns
-        -------
-        dict
-        """
-        return self._profile
-
-    @property
-    def transform(self):
-        """rasterio.Affine: Affine transform for exclusion layer. """
-        return Affine(*self.profile["transform"])
-
-    @property
-    def arr_shape(self):
-        """Rasterize array shape.
-
-        Returns
-        -------
-        tuple
-        """
-        return self._shape
-
-    @property
-    def inclusions(self):
-        """Flag indicating whether or not the output raster represents
-        inclusion values.
-
-        Returns
-        -------
-        bool
-        """
-        return self._scale_factor > 1
-
-    def _no_exclusions_array(self, multiplier=1, window=None):
-        """Get an array of the correct shape representing no exclusions.
-
-        The array contains all zeros, and a new one is created
-        for every function call.
-
-        Parameters
-        ----------
-        multiplier : int, optional
-            Integer multiplier value used to scale up the dimensions of
-            the array exclusions array (e.g. multiplier of 3 turns an
-            array of shape (10, 20) into an array of shape (30, 60)).
-        window : :cls:`rasterio.windows.Window`
-            A ``rasterio`` window defining the area of the raster. Can
-            be used to speed up computation and decrease memory
-            requirements if features are localized to a small portion of
-            the raster array.
-
-        Returns
-        -------
-        np.array
-            Array of zeros representing no exclusions.
-        """
-        if window is None:
-            shape = tuple(x * multiplier for x in self.arr_shape[1:])
-        else:
-            shape = (window.height * multiplier, window.width * multiplier)
-        return np.zeros(shape, dtype='uint8')
-
-    def rasterize(self, shapes, window=None):
-        """Convert geometries into exclusions array.
-
-        Parameters
-        ----------
-        shapes : list, optional
-            List of geometries to rasterize (i.e. list(gdf["geometry"])).
-            If `None` or empty list, returns array of zeros.
-        window : :obj:`rasterio.windows.Window`
-            A ``rasterio`` window defining the area of the raster. Can
-            be used to speed up computation and decrease memory
-            requirements if features are localized to a small portion of
-            the raster array.
-
-        Returns
-        -------
-        arr : ndarray
-            Rasterized array of shapes.
-        """
-
-        shapes = shapes or []
-        shapes = [(geom, 1) for geom in shapes if geom is not None]
-
-        if self.inclusions:
-            return self._rasterize_to_weights(shapes, window)
-
-        return self._rasterize_to_mask(shapes, window)
-
-    def _rasterize_to_weights(self, shapes, window):
-        """Rasterize features to weights using a high-resolution array."""
-
-        if not shapes:
-            return ((1 - self._no_exclusions_array(window=window))
-                    .astype(np.float32))
-
-        hr_arr = self._no_exclusions_array(multiplier=self._scale_factor,
-                                           window=window)
-        transform = self._window_transform(window)
-        transform *= transform.scale(1 / self._scale_factor)
-
-        rio_features.rasterize(shapes=shapes, out=hr_arr, fill=0,
-                               transform=transform)
-
-        arr = self._aggregate_high_res(hr_arr, window)
-        return 1 - (arr / self._scale_factor ** 2)
-
-    def _rasterize_to_mask(self, shapes, window):
-        """Rasterize features with to an exclusion mask."""
-
-        arr = self._no_exclusions_array(window=window)
-        if shapes:
-            transform = self._window_transform(window)
-            rio_features.rasterize(shapes=shapes, out=arr, fill=0,
-                                   transform=transform)
-
-        return arr
-
-    def _aggregate_high_res(self, hr_arr, window):
-        """Aggregate the high resolution exclusions array to output shape. """
-
-        arr = self._no_exclusions_array(window=window).astype(np.float32)
-        for i, j in product(range(self._scale_factor),
-                            range(self._scale_factor)):
-            arr += hr_arr[i::self._scale_factor, j::self._scale_factor]
-        return arr
-
-    def _window_transform(self, window):
-        """Calculate the transform for a given window, if any. """
-        if window is None:
-            return deepcopy(self.transform)
-        return windows.transform(window, self.transform)
 
 
 class AbstractBaseSetbacks(AbstractBaseExclusionsMerger):
@@ -396,7 +211,7 @@ class AbstractBaseSetbacks(AbstractBaseExclusionsMerger):
         features = FEATURE_FILTERS[features_filter_type](features, county)
         features = BUFFERS[buffer_type](features, regulation_value)
 
-        return _rasterize_within_window(features, feature_bounds, rasterizer)
+        return rasterizer.rasterize_within_window(features, feature_bounds)
 
     def compute_generic_exclusions(self, max_workers=None):
         """Compute generic setbacks.
@@ -537,40 +352,6 @@ class AbstractBaseSetbacks(AbstractBaseExclusionsMerger):
         raise NotImplementedError
 
 
-def _parse_excl_properties(excl_fpath, hsds=False):
-    """Parse shape, chunk size, and profile from exclusions file.
-
-    Parameters
-    ----------
-    excl_fpath : str
-        Path to .h5 file containing exclusion layers, will also be
-        the location of any new setback layers
-    hsds : bool, optional
-        Boolean flag to use h5pyd to handle .h5 'files' hosted on
-        AWS behind HSDS. By default `False`.
-
-    Returns
-    -------
-    shape : tuple
-        Shape of exclusions datasets
-    profile : str
-        GeoTiff profile for exclusions datasets
-    """
-    with ExclusionLayers(excl_fpath, hsds=hsds) as exc:
-        dset_shape = exc.shape
-        profile = exc.profile
-
-    if len(dset_shape) < 3:
-        dset_shape = (1, ) + dset_shape
-
-    logger.debug('Exclusions properties:\n'
-                 'shape : {}\n'
-                 'profile : {}\n'
-                 .format(dset_shape, profile))
-
-    return dset_shape, profile
-
-
 def _load_features(features_ids, features_fp, col, crs):
     """Load the `features_ids` from the `features_fp`. """
     ids = ",".join(map(str, features_ids))
@@ -603,7 +384,7 @@ def _compute_exclusions(features_ids, features_fp, col, crs, buffer_type,
     if setbacks is None:
         return None, None
 
-    return _rasterize_within_window(setbacks, feature_bounds, rasterizer)
+    return rasterizer.rasterize_within_window(setbacks, feature_bounds)
 
 
 def _load_and_buffer(features_ids, features_fp, col, crs, buffer_type,
@@ -631,32 +412,3 @@ def _buffered_feature_bounds(features, rasterizer, regulation_value):
     buffer_len = max(abs(rasterizer.transform.a), abs(rasterizer.transform.e))
     bound_buffer = regulation_value * 2 + buffer_len
     return features.buffer(bound_buffer).total_bounds
-
-
-def _rasterize_within_window(features, bounds, rasterizer):
-    """Rasterize the features using the geoseries bounding box as a window. """
-    window = _cropped_window(bounds, rasterizer.transform,
-                             rasterizer.arr_shape[1:])
-    logger.debug(f"Rasterizing {len(features)} features using {window!r}")
-    exclusions = rasterizer.rasterize(features, window=window)
-    log_mem(logger)
-    logger.debug(f"Exclusion mem size: {exclusions.nbytes / 1048576:.2f}MB")
-    return exclusions, window.toslices()
-
-
-def _cropped_window(bounds, raster_transform, shape):
-    """Calculate the raster array window corresponding to the bounding box."""
-    left, bottom, right, top = bounds
-
-    rows, cols = transform.rowcol(raster_transform,
-                                  [left, right, right, left],
-                                  [top, top, bottom, bottom],
-                                  op=float)
-
-    row_start = max(floor(min(rows)), 0)
-    col_start = max(floor(min(cols)), 0)
-    row_stop = min(ceil(max(rows)), shape[0])
-    col_stop = min(ceil(max(cols)), shape[1])
-    return windows.Window(col_off=col_start, row_off=row_start,
-                          width=max(col_stop - col_start, 0),
-                          height=max(row_stop - row_start, 0))
